+++ app/src/components/Header/Header.js
@@ -1,93 +1,7 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
+import { withRouter } from 'react-router-dom';
+import { OptionsTop, OptionsLeft } from './constHeader.js';
 import './Header.css'
-
-
-const top = [[
-  {id:1, name: 'Iniciar Sesión', link: '/login', icon: 'fas fa-sign-in-alt'},
-  {id:2, name: 'Ayuda', link: '/about', icon: 'far fa-question-circle'}
-], [
-  {id:3, name: 'Cerrar Sesión', link: '/', icon: 'fas fa-sign-out-alt'},
-  {id:4, name: 'Ayuda', link: '/about', icon: 'far fa-question-circle'},
-]];
-const left = [[
-  {id:5, name: 'Inicio', link: '/', icon: 'fas fa-home'},
-  {id:6, name: 'Iniciar Sesión', link: '/login', icon: 'fas fa-sign-in-alt'}
-],[
-  {id:7, name: 'Inicio', link: '/', icon: 'fas fa-home'},
-  {id:8, name: 'Mis Contraseñas', link: '/', icon: 'fas fa-key'},
-  {id:9, name: 'Ajustes', link: '/', icon: 'fas fa-cog'},
-  {id:10, name: 'Cerrar sesión', link: '/', icon: 'fas fa-sign-out-alt'}
-]];
-
-const OptionsTop = props => {
-  if (props.mobile) {
-    return null;
-  }
-  let user = 0, log = null;
-  if (props.user){
-    /*log = (
-      <li key={user} className="nav-top-item nav-text">
-        {props.user.name}
-      </li>
-    )*/
-    user = 1
-  }
-  return (
-    <div className="collapse">
-      <ul>
-      {log}
-      {props.opts[user].map((opt) =>
-        <li key={opt.id}>
-          <Link to={opt.link} className="nav-top-item">
-            <span className={opt.icon}/> {opt.name}
-          </Link>
-        </li>
-      )}
-      </ul>
-    </div>
-  );
-};
-
-const OptionsLeft = props => {
-  let cls = "nav-vertical"
-  if (props.mobile) {
-    cls += " responsive";
-  }
-  let user = 0;//, log = null;
-  if (props.user){
-    user = 1
-  }
-  return(
-    <div className={cls}>
-      <ul>
-      {props.opts[user].map((opt) =>
-        <li key={opt.id} className="nav-left-item">
-          <Link to={opt.link}>
-            <span className={opt.icon}/>
-            <i>{opt.name}</i>
-          </Link>
-        </li>
-      )}
-      </ul>
-      <div className="spacer"/>
-      <ul>
-        <li className="nav-left-item">
-          <a href="https://www.google.com/">
-            <span className="fas fa-book"/>
-            <i>Documentación</i>
-          </a>
-        </li>
-        <li className="nav-left-item">
-          <a href="https://www.google.com/">
-            <span className="fas fa-info-circle"/>
-            <i>About</i>
-          </a>
-        </li>
-      </ul>
-    </div>
-  );
-};
 
 function toogle() {
   // console.log("onClick");
@@ -185,11 +99,14 @@ class Header extends React.Component {
 
   render(){
     //console.log("render " + this.state.mobile);
+    //console.log(this.props.location);
     //let user = null;
-    let user = {name: "ads"}
-    const tp = <OptionsTop mobile={this.state.mobile} user={user} opts={top}/>;
-    const lf = <OptionsLeft mobile={this.state.mobile} user={user} opts={left}/>
-    if (!this.state.mobile){
+    let user = {name: "ads"}, loc = this.props.location.pathname;
+    const tp = <OptionsTop mobile={this.state.mobile}
+                    user={user} currentlocation={loc}/>;
+    const lf = <OptionsLeft mobile={this.state.mobile}
+                    user={user} currentlocation={loc}/>
+    /*if (!this.state.mobile){
       return(
         <Static top={tp} left={lf}/>
       );
@@ -198,13 +115,10 @@ class Header extends React.Component {
         <Static top={tp} left={lf}/>
       );
     }
-    return null;
+    return null;*/
+    return (<Static top={tp} left={lf}/>);
   }
 }
 
-//// TODO:  -collapse con cuerpo de App
-////        -spacer entre listas lef-navbar
-////        -active? => state?
-////        -
-////        -
-export default Header;
+
+export default Header = withRouter(Header);
